#28 Increases top padding for posts container.
Increases the top padding of the posts container to make mobile rendering easier to read, and to maintain consistency with all the other spacing shown on the site. Also removes all post header top margins and paddings to make its spacing match that of the pure text elements. This should make the mobile site more coherent and cleaner to look at.