#29 Increases line height for mobile blog links, hides navbar on mobile.
Increases the line height of the blog post links on mobile screens so that their dotten lines don't overlap. Also noticed that the navigation bar was visible at around 750px, the same time that the mobile navigation header is visible. This caused overlapping with the two elements! I updated the breakpoint for the main navigation bar so that this doesn't happen.