#27 Hides title in titlebar on mobile.
Hides the title on the title bar element on mobile since it's redundant. Also removes the title bar element in general when there are no links for the user to click on (no need for the extra element when it literally does nothing). Finally, removes the top padding and margin on text elements that was there to compensate for the size of the title bar element.